scripts/pyproject.toml
[project]
name = "rime-configuration"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "WithdewHua", email = "i@10101.io" }
]
dependencies = [
    "requests>=2.31.0",
    "pyyaml>=6.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["."]
